+++ .vscode/settings.json
@@ -19,6 +19,7 @@
     "onkeydown",
     "onlcick",
     "radiogroup",
+    "Routable",
     "Slottable",
     "structopt",
     "tabindex",
