+++ roving_focus/src/hooks/mod.rs
@@ -1,3 +1,2 @@
 pub mod use_children_as_html_collection;
-pub mod use_keydown;
 pub mod use_roving_iterator;
