--- slot/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "slot"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-yew = "0.21.0"
-log = "0.4.21"
