+++ storybook/src/bin/server.rs
@@ -1,23 +1,33 @@
 #[macro_use]
 extern crate rocket;
 
-use std::path::PathBuf;
+use std::{collections::HashMap, path::PathBuf};
 
 use clap::Parser;
 use rocket::response::stream::TextStream;
 use rocket::{response::content::RawHtml, State};
-use storybook::App;
+use storybook::{App, AppProps};
 
 struct IndexHtml {
     pub head: String,
     pub body: String,
 }
 
-#[get("/")]
-async fn index(index_html: &State<IndexHtml>) -> RawHtml<TextStream![String]> {
-    let renderer = yew::ServerRenderer::<App>::new();
+#[get("/<path..>?<query..>", rank = 11)]
+async fn index(
+    index_html: &State<IndexHtml>,
+    path: PathBuf,
+    query: Option<HashMap<String, String>>,
+) -> RawHtml<TextStream![String]> {
+    let renderer = yew::ServerRenderer::<App>::with_props(move || AppProps {
+        path: path.to_string_lossy().to_string().into(),
+        query: query.unwrap_or_default().into(),
+    });
+
     let index_html_head = index_html.head.clone();
     let index_html_body = index_html.body.clone();
+
+    log::info!("index_html_head: {}", index_html_head);
 
     RawHtml(TextStream! {
         yield index_html_head;
