+++ storybook/src/pages/mod.rs
@@ -1,0 +1,3 @@
+mod storybook;
+
+pub use storybook::StorybookPage;
