+++ switch/Cargo.toml
@@ -8,3 +8,4 @@ edition = "2021"
 [dependencies]
 yew = "0.21.0"
 utils = { path = "../utils" }
+log = "0.4.22"
