+++ utils/src/hooks/mod.rs
@@ -1,1 +1,5 @@
 pub mod use_controllable_state;
+pub mod use_keydown;
+
+pub use use_controllable_state::use_controllable_state;
+pub use use_keydown::use_keydown;
